feature/relatorios: cria visoes financeira operacional e gerencial
